+++ attributes/default.rb
@@ -5,7 +5,7 @@ default[:etcd][:install_method] = 'binary'
 default[:etcd][:seed_node] = nil
 
 # if you wrap this cookbook you should use your wrappers cook name here
-default[:etcd][:search_cook] = 'etcd'
+default[:etcd][:search_cook] = 'etcd\:\:cluster'
 
 # set to false if you don't want environment scoped searching
 default[:etcd][:env_scope] = true