+++ test/spec/spec_helper.rb
@@ -1,18 +1,11 @@
-$: << File.expand_path(File.dirname(__FILE__) + "../../")
-
-if ENV["COVERAGE"]
-  require "simplecov"
-
-  SimpleCov.start do
-    add_filter "spec"
-  end
-end
-
+cook_dir = File.expand_path(File.dirname(__FILE__) + "../../../.cooks")
 require "chefspec"
-require 'chefspec/berkshelf'
+require 'chefspec/server'
 
 RSpec.configure do |config|
-  config.platform = 'centos'
-  config.version = '6.4'
+  config.platform = 'ubuntu'
+  config.version = '12.04'
   config.color = true
+  config.cookbook_path = cook_dir
 end
+